// src/types/spotify_api.ts
export type ISO3166_1_Alpha_2 =
    | 'AF' | 'AX' | 'AL' | 'DZ' | 'AS' | 'AD' | 'AO' | 'AI' | 'AQ' | 'AG'
    | 'AR' | 'AM' | 'AW' | 'AU' | 'AT' | 'AZ' | 'BS' | 'BH' | 'BD' | 'BB'
    | 'BY' | 'BE' | 'BZ' | 'BJ' | 'BM' | 'BT' | 'BO' | 'BQ' | 'BA' | 'BW'
    | 'BV' | 'BR' | 'IO' | 'BN' | 'BG' | 'BF' | 'BI' | 'KH' | 'CM' | 'CA'
    | 'CV' | 'KY' | 'CF' | 'TD' | 'CL' | 'CN' | 'CX' | 'CC' | 'CO' | 'KM'
    | 'CG' | 'CD' | 'CK' | 'CR' | 'CI' | 'HR' | 'CU' | 'CW' | 'CY' | 'CZ'
    | 'DK' | 'DJ' | 'DM' | 'DO' | 'EC' | 'EG' | 'SV' | 'GQ' | 'ER' | 'EE'
    | 'SZ' | 'ET' | 'FK' | 'FO' | 'FJ' | 'FI' | 'FR' | 'GF' | 'PF' | 'TF'
    | 'GA' | 'GM' | 'GE' | 'DE' | 'GH' | 'GI' | 'GR' | 'GL' | 'GD' | 'GP'
    | 'GU' | 'GT' | 'GG' | 'GN' | 'GW' | 'GY' | 'HT' | 'HM' | 'VA' | 'HN'
    | 'HK' | 'HU' | 'IS' | 'IN' | 'ID' | 'IR' | 'IQ' | 'IE' | 'IM' | 'IL'
    | 'IT' | 'JM' | 'JP' | 'JE' | 'JO' | 'KZ' | 'KE' | 'KI' | 'KP' | 'KR'
    | 'KW' | 'KG' | 'LA' | 'LV' | 'LB' | 'LS' | 'LR' | 'LY' | 'LI' | 'LT'
    | 'LU' | 'MO' | 'MG' | 'MW' | 'MY' | 'MV' | 'ML' | 'MT' | 'MH' | 'MQ'
    | 'MR' | 'MU' | 'YT' | 'MX' | 'FM' | 'MD' | 'MC' | 'MN' | 'ME' | 'MS'
    | 'MA' | 'MZ' | 'MM' | 'NA' | 'NR' | 'NP' | 'NL' | 'NC' | 'NZ' | 'NI'
    | 'NE' | 'NG' | 'NU' | 'NF' | 'MK' | 'MP' | 'NO' | 'OM' | 'PK' | 'PW'
    | 'PS' | 'PA' | 'PG' | 'PY' | 'PE' | 'PH' | 'PN' | 'PL' | 'PT' | 'PR'
    | 'QA' | 'RE' | 'RO' | 'RU' | 'RW' | 'BL' | 'SH' | 'KN' | 'LC' | 'MF'
    | 'PM' | 'VC' | 'WS' | 'SM' | 'ST' | 'SA' | 'SN' | 'RS' | 'SC' | 'SL'
    | 'SG' | 'SX' | 'SK' | 'SI' | 'SB' | 'SO' | 'ZA' | 'GS' | 'SS' | 'ES'
    | 'LK' | 'SD' | 'SR' | 'SJ' | 'SE' | 'CH' | 'SY' | 'TW' | 'TJ' | 'TZ'
    | 'TH' | 'TL' | 'TG' | 'TK' | 'TO' | 'TT' | 'TN' | 'TR' | 'TM' | 'TC'
    | 'TV' | 'UG' | 'UA' | 'AE' | 'GB' | 'US' | 'UM' | 'UY' | 'UZ' | 'VU'
    | 'VE' | 'VN' | 'VG' | 'VI' | 'WF' | 'EH' | 'YE' | 'ZM' | 'ZW';

export type Locale = `${ISO639_1}_${ISO3166_1_Alpha_2}`;

export type ExternalUrls = {
    spotify: string;
}

export type AlbumType = 'album' | 'single' | 'compilation';

export type DatePrecision = 'day' | 'month' | 'year';

export type RestrictionReason = 'market' | 'product' | 'explicit';

export type ResourceType = "album" | "artist" | "playlist" | "track" | "show" | "episode" | "audiobook";

export type Restrictions = {
    reason: RestrictionReason;
};

export type Image = {
    url: string;
    height: number;
    weight: number;
}

export type CopyrightObject = {
    text: string;
    type: 'C' | 'P';
}

type ExternalIds = {
    isrc: string;
    ean: string;
    upc: string;
}

export interface Paged<T> {
    href: string;
    limit: number;
    next?: string;
    offset: number;
    previous?: string;
    total: number;
    items: Array<T | null>;
}

export interface Album {
    album_type: AlbumType;
    total_tracks: number;
    available_markets: Array<ISO3166_1_Alpha_2>;
    external_urls: ExternalUrls;
    href: string;
    id: string;
    images: Array<Image>;
    name: string;
    release_date: string;
    release_date_precision: DatePrecision;
    restrictions: Restrictions;
    type: 'album';
    uri: string;
    artists: Array<SimplifiedArtist>;
    tracks: Paged<SimplifiedTrack>;
    copyrights: Array<CopyrightObject>;
    external_ids: ExternalIds;
    label: string;
    popularity: number;
}

export interface SimplifiedAlbum extends Omit<Album, 'tracks' | 'copyrights' | 'external_ids' | 'label' | 'popularity'> { }

export interface Artist {
    external_urls: ExternalUrls;
    followers: {
        href: string;
        total: number;
    };
    genres: Array<string>;
    href: string;
    id: string;
    images: Array<Image>;
    name: string;
    popularity: number;
    type: string;
    uri: string;
}

export interface SimplifiedArtist extends Omit<Artist, 'followers' | 'images' | 'popularity' | 'genres'> { }
// ...
export interface Track {
    album: SimplifiedAlbum;
    artists: Array<SimplifiedArtist>;
    available_markets: Array<ISO3166_1_Alpha_2>;
    disc_number: number;
    duration_ms: number;
    explicit: boolean;
    external_ids: ExternalIds;
    external_urls: ExternalUrls;
    href: string;
    id: string;
    is_playable: boolean;
    linked_from?: {};
    restrictions: Restrictions;
    name: string;
    popularity: number;
    track_number: number;
    type: 'track';
    uri: string;
    is_local: boolean;
}

export interface SimplifiedTrack extends Omit<Track, 'album' | 'external_ids' | 'popularity'> { }